requirements.txt
torch>=2.0.0
transformers>=4.30.0
streamlit>=1.28.0
pyyaml>=6.0
numpy>=1.24.0
steering-vectors
accelerate